pyproject.toml
[tool.poetry]
name = "fastapi-server"
version = "0.1.0"
description = "A simple and modularized boilerplate"
authors = ["jose <jose.s.contacto@gmail.com>"]
readme = "README.md"
packages = [{include = "fastapi_server"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.105.0"
bcrypt = "^4.1.2"
asyncio = "^3.4.3"
uvicorn = "^0.24.0.post1"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest-watch = "^4.2.0"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
